requirements.txt
Django==1.8.5
djangorestframework==3.3.2
django-filter==0.12.0
dj-database-url==0.4.0
psycopg2==2.6.1
gunicorn==19.4.5
Markdown==2.6.5
whitenoise==2.0.6

# Dicom
pydicom==0.9.9
pillow==2.9.0
numpy==1.9.2
